refactor: move probe detector modules into subpackage
